requirements.txt
Flask>=2.3.2
PyPDF2>=3.0.0
requests>=2.31.0
ollama
openai 
pdfplumber
python-docx
markdown
pywebview
pyinstaller
